Tutorial and mana node bug fixes (#62)
- Fix tutorial steps being changed in an unexpected way after the
tutorial skip override, by making it so a call to /tutorial/update_step
will not update the tutorial step if the request value is lower than the
db value. This was happening as the endpoint was called after the
tutorial skip
- This was also leading to the WP tutorial being triggered and causing a
softlock before an upgrade was attempted because the print at index = 0
in new savefiles is maxed out. Adding a junk print (Mana Fount) before
the other prints will fix this when we later enable this part of the
tutorial.
- Adjust unawaited async void call in CharaController, hopefully closing
#61
- Fix MC50 bonus not being applied with omnicite, and add correct stat
check to omnicite test, closing #60
- Change coins to be a long in the db (matching the generated type) so
that over 2 billion coins can be owned. Migration also added for this.

+++ DragaliaAPI.Database/Entities/DbPlayerUserData.cs
@@ -35,7 +35,7 @@ public class DbPlayerUserData : IDbHasAccountId
     /// <summary>
     /// The player's rupie balance.
     /// </summary>
-    public int Coin { get; set; }
+    public long Coin { get; set; }
 
     public int MaxDragonQuantity { get; set; }
